requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
redis>=4.0.0
supabase>=0.7.0
torch>=2.7.0
pytest~=8.3.5
pandas~=2.2.3
numpy~=1.26.4
matplotlib~=3.10.0
plotly~=6.0.0
scikit-learn~=1.6.1
seaborn~=0.13.2
tqdm~=4.67.1
scipy~=1.15.2
python-jose~=3.4.0